Support new decorators proposal as implemented in Babel 7
* Moves @property tests to decorators_test.ts
* Adds a Babel build step to compile decorators_test.ts to decorators-babel_test.js
* Updates decorators to detect if they’re called in legacy or modern decorator mode

Fixes #156
