+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "plugins":[
+    ["@babel/plugin-proposal-decorators", {"decoratorsBeforeExport": true}],
+    "@babel/plugin-proposal-class-properties",
+    "@babel/plugin-transform-typescript"]
+}
