+++ test/runner.html
@@ -19,5 +19,6 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
   <script type="module" src="lit-element_test.js"></script>
   <script type="module" src="lit-element_styling_test.js"></script>
   <script type="module" src="lib/decorators_test.js"></script>
+  <script type="module" src="lib/decorators-babel_test.js"></script>
   <script type="module" src="lib/updating-element_test.js"></script>
 </body></html>
